requirements.txt
streamlit>=1.33,<2
ultralytics>=8.2.0,<9
pillow>=10,<12
onnxruntime>=1.17,<2
opencv-python-headless>=4.8,<5
numpy<2
